[🚀 사이클2 - 미션 (예약 변경/취소와 에러 처리)] 카야 미션 제출합니다. (#505)
* feat: 예약 시간 및 예약 테마 존재 여부 확인 메서드 추가

* feat: 예약이 존재할 때 예약시간 변경 금지 조건 추가

* feat: 예약이 존재할 때 테마 변경 금지 조건 추가

* feat: hard delete에서 delete 막음으로 변경

* feat: 에러 반환 시 상태 코드 반환하도록 추가

* feat: 예약에서 ErrorCode 사용하도록 변경

* feat: 시간에서 ErrorCode 사용하도록 변경

* feat: 테마에서 ErrorCode 사용하도록 변경

* feat: 테스트 코드도 에러 코드를 반환하는 코드로 변경

* feat: 에러 발생 시 ResponseEntity에 에러 코드도 담아서 전달

* feat: 메소드명 오타 변경

* feat: 내 이름으로 예약 리스트 조회 기능 생성

* feat: 내 이름으로 예약 업데이트 기능 생성

* feat: 내 이름으로 예약 삭제 기능 생성

* feat: 에러 코드 추가

* chore: import 문 추가

* feat: jdbc repository 업데이트

* test: 테스트 코드 추가

* feat: 프론트 화면 수정(내 예약보기, 내 예약 수정, 삭제 구현)

* feat: reservation 중복 예약 시 db에서 에러 반환하고, 이를 도메인 에러로 변환

* test: 예약 동시성 테스트 추가

* feat: 예약 디테일을 가져오는 메서드 추가

* refactor: 예약 update/delete 시 ReservationDetail을 가져와서 검증하도록 변경

* refactor: reservation에 isOwner 메서드 추가

* refactor: update 시 updatedRowCount 확인하기

* test: FakeREservationRepository에 findDetailById 구현하기

+++ src/main/java/roomescape/reservation/application/service/ReservationService.java
@@ -7,9 +7,11 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.global.RoomEscapeException;
 import roomescape.reservation.application.dto.ReservationCreateCommand;
 import roomescape.reservation.application.dto.ReservationQueryResult;
-import roomescape.reservation.application.exception.ReservationException;
+import roomescape.reservation.application.dto.ReservationUpdateCommand;
+import roomescape.reservation.application.exception.ReservationErrorCode;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.repository.ReservationDetail;
 import roomescape.reservation.domain.repository.ReservationRepository;
@@ -35,6 +37,13 @@ public class ReservationService {
                 .toList();
     }
 
+    @Transactional(readOnly = true)
+    public List<ReservationQueryResult> findAllByName(String name) {
+        return reservationRepository.findByName(name).stream()
+                .map(this::toQueryResult)
+                .toList();
+    }
+
     public ReservationQueryResult save(ReservationCreateCommand request, LocalDateTime currentDateTime) {
         ReservationTimeQueryResult timeQueryResult = timeService.findById(request.timeId());
         validateReservationDateTime(request.date(), timeQueryResult.startAt(), currentDateTime);
@@ -46,16 +55,32 @@ public class ReservationService {
         return ReservationQueryResult.from(reservationRepository.save(reservation), themeQueryResult, timeQueryResult);
     }
 
-    public int delete(Long id) {
+    public ReservationQueryResult update(ReservationUpdateCommand request, LocalDateTime currentDateTime) {
+        ReservationDetail reservationDetail = getReservationDetail(request.id());
+        Reservation reservation = toReservation(reservationDetail);
+        validateOwner(request.name(), reservation);
+        validateReservationNotPast(reservationDetail, currentDateTime);
+
+        ReservationTimeQueryResult timeQueryResult = timeService.findById(request.timeId());
+        validateReservationDateTime(request.date(), timeQueryResult.startAt(), currentDateTime);
+        validateDuplicateReservation(request, reservation);
+
+        Reservation updatedReservation = reservation.update(request.date(), request.timeId());
+        Reservation savedReservation = reservationRepository.update(updatedReservation);
+        return toQueryResult(savedReservation);
+    }
+
+    public int delete(Long id, String name, LocalDateTime currentDateTime) {
+        ReservationDetail reservationDetail = getReservationDetail(id);
+        Reservation reservation = toReservation(reservationDetail);
+        validateOwner(name, reservation);
+        validateReservationNotPast(reservationDetail, currentDateTime);
         return reservationRepository.delete(id);
     }
 
-    private void validateReservationDateTime(LocalDate date, LocalTime startAt, LocalDateTime currentDateTime) {
-        LocalDateTime triedDateTime = LocalDateTime.of(date, startAt);
-
-        if (triedDateTime.isBefore(currentDateTime)) {
-            throw new ReservationException("현재 시간보다 이전 시간으로 예약을 할 수 없습니다.");
-        }
+    private ReservationDetail getReservationDetail(Long id) {
+        return reservationRepository.findDetailById(id)
+                .orElseThrow(() -> new RoomEscapeException(ReservationErrorCode.RESERVATION_NOT_FOUND));
     }
 
     private void validateDuplicateReservation(ReservationCreateCommand request) {
@@ -64,7 +89,57 @@ public class ReservationService {
                 request.timeId()
         );
         if (existsByDateAndTime) {
-            throw new ReservationException("이미 해당 날짜와 시간에 예약이 존재합니다.");
+            throw new RoomEscapeException(ReservationErrorCode.DUPLICATE_RESERVATION);
         }
+    }
+
+    private void validateDuplicateReservation(ReservationUpdateCommand request, Reservation reservation) {
+        Boolean existsByDateAndTime = reservationRepository.existsByDateAndThemeAndTimeExcludingId(
+                request.date(),
+                reservation.getThemeId(),
+                request.timeId(),
+                reservation.getId()
+        );
+        if (existsByDateAndTime) {
+            throw new RoomEscapeException(ReservationErrorCode.DUPLICATE_RESERVATION);
+        }
+    }
+
+    private void validateOwner(String name, Reservation reservation) {
+        if (!reservation.isOwner(name)) {
+            throw new RoomEscapeException(ReservationErrorCode.FORBIDDEN_RESERVATION_ACCESS);
+        }
+    }
+
+    private void validateReservationDateTime(LocalDate date, LocalTime startAt, LocalDateTime currentDateTime) {
+        LocalDateTime triedDateTime = LocalDateTime.of(date, startAt);
+
+        if (triedDateTime.isBefore(currentDateTime)) {
+            throw new RoomEscapeException(ReservationErrorCode.PAST_RESERVATION_TIME);
+        }
+    }
+
+    private void validateReservationNotPast(ReservationDetail reservationDetail, LocalDateTime currentDateTime) {
+        LocalDateTime reservationDateTime = LocalDateTime.of(reservationDetail.date(), reservationDetail.startAt());
+
+        if (reservationDateTime.isBefore(currentDateTime)) {
+            throw new RoomEscapeException(ReservationErrorCode.PAST_RESERVATION_MODIFICATION);
+        }
+    }
+
+    private ReservationQueryResult toQueryResult(Reservation reservation) {
+        ThemeQueryResult themeQueryResult = themeService.findById(reservation.getThemeId());
+        ReservationTimeQueryResult timeQueryResult = timeService.findById(reservation.getTimeId());
+        return ReservationQueryResult.from(reservation, themeQueryResult, timeQueryResult);
+    }
+
+    private Reservation toReservation(ReservationDetail reservationDetail) {
+        return Reservation.builder()
+                .id(reservationDetail.reservationId())
+                .name(reservationDetail.username())
+                .date(reservationDetail.date())
+                .themeId(reservationDetail.themeId())
+                .timeId(reservationDetail.timeId())
+                .build();
     }
 }
